added:Username Validation

+++ routes/ideas.js
@@ -45,17 +45,24 @@ router.post("/", async (req, res) => {
 //PUT Update idea
 router.put("/:id", async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag
-        }
-      },
-      { new: true }
-    );
-    res.json({ success: true, data: updatedIdea });
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the username
+    if(idea.username === req.body.username){
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag
+          }
+        },
+        { new: true }
+      );
+      return res.json({ success: true, data: updatedIdea });
+    }
+    // Usernames do not match
+    res.status(403).json({ success:false, error:'You are not authorized to update this resource!' });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success:false, error:'Something went wrong!' });
@@ -65,8 +72,15 @@ router.put("/:id", async (req, res) => {
 //DELETE An idea
 router.delete("/:id", async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.status(200).json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the username
+    if(idea.username === req.body.username){
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.status(200).json({ success: true, data: {} });
+    }
+    // Usernames do not match
+    res.status(403).json({ success:false, error:'You are not authorized to delete this resource!' });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success:false, error:'Something went wrong!' });
